services/upscaler/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
numpy<2
torch==2.1.0
torchvision==0.16.0
realesrgan
basicsr
facexlib
gfpgan